Meilleur structuration du json dispatcher
