+++ json_dispatcher/error_mapping.inc.php
@@ -1,0 +1,11 @@
+<?php
+
+
+$error_mapping = array(
+	'ServiceNotFound' => 404,
+	'ServiceMethodNotFound' => 404,
+	'ServiceMethodForbidden' => 403,
+	'ServiceMissingMethodArgument' => 400,
+);
+
+
