+++ json_dispatcher/services.inc.php
@@ -1,0 +1,9 @@
+<?php
+
+$services = array(
+	'POSS' => 'services/POSS.service.php',
+	'POSS2' => 'services/POSS2.service.php',
+	'AADMIN' => 'services/AADMIN.service.php',
+	'MADMIN' => 'services/MADMIN.service.php',
+	'STATS' => 'services/STATS.service.php',
+);
